Automate joining new channels when archiving channels
In order to read a channel's history (and probably also to archive it),
we need to be a member of that channel. We want to do that so we can
later determine whether a channel is inactive, and so safe to archive.

Slack imposes an upper limit of 999 (or maybe 1000) channels in a single
list query of "conversations" (which includes the option of private
channels, DMs, and group DMs, as well as public channels - we only care
about the last), but paginates them. We don't yet support that
pagination, so we only use whatever comes in the first request.