Only archive channels inactive channels
We consider a channel stale if, in the last 30 days, it has had no
visible activity other than bot messages or users joining or leaving.

For now, we're testing that our private archive method is being called,
but in the long run, we'll want to be testing agains the Slack client
itself, once we actually archive the channels.