Add support for marking channels as no archive
We want to be able to keep quiet channels running without archiving
them. This is useful for channels that have occasional bursts of
activity, or channels which are mostly for monitoring bot output.

We do this by providing support for a magic string that can be put into
the channel's description or topic to tell the bot to ignore it.