Include channels covered by the rules with disabled defaults
We want to monitor the channels with rules. Normally, all channels are
covered by the rules, as we monitor every channel (that doesn't have one
of the criteria that excludes it), but when the default rules are
disabled, channel monitoring is opt in, so we need to check to see if a
channel matches before joining and monitoring it.