Make ArchiveChannels require instancing
This allows us to cache things indefinitely, knowing that we'll create a
new instance each time we want to run it.

It's ok to effectively take a snapshot of the Slack workspace each time
we run the archive job, so we can save on the number of API calls for
repeated data by memoizing them. The main caveat is that we flush the
cache before we start a run.